requirements.txt
streamlit==1.31.0
openai==0.28.1
pillow==9.5.0
requests==2.28.0
gtts==2.3.0
python-dotenv==1.0.0
# numpy와 pandas는 streamlit에 이미 포함되어 있음